requirements.txt
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.1
google-api-core==2.24.0
google-api-python-client==2.165.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.22.0
setuptools==78.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.45.1
wsproto==1.2.0
yt-dlp==2025.1.15
